agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Base image for kubeguardian:agent"
requires-python = ">=3.12.4"
dependencies = [
    "google-adk",
    "pydantic",
    "uvicorn",
    "fastapi",
    "Flask",
    "python-dotenv",    
    "bcrypt",          
    "requests",         
    "SQLAlchemy",
    "psycopg2-binary"
]
